requirements.txt
annotated-types==0.6.0
anyio==3.7.1
certifi==2023.7.22
charset-normalizer==3.3.1
click==8.1.7
exceptiongroup==1.1.3
fastapi==0.104.0
h11==0.14.0
idna==3.4
mysqlclient==2.2.0
pydantic==2.4.2
pydantic_core==2.10.1
python-dotenv==1.0.0
python-multipart==0.0.6
requests==2.31.0
PyJWT
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.8.0
urllib3==2.0.7
uvicorn==0.23.2